matlab: add quantization error to get_scaled_csi
For strong links, CSI quantization error can dominate noise. Thus, it's
important to take this into account when computing the noise floor of
the H matrix.

The quantization error is roughly 36 dB below RSS, but calculated in a
way that reflects the scaling of a particular CSI matrix.

This makes CSI much more accurate for strong links, improves algorithmic
performance, and appears to get rid of the nasty 'Inf' Effective SNR
values that complicate code.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>