matlab: fix quantization error calculation
I realized that the CSI-normalized error is only true if the receiver is
using the CSI itself to do its computation. But the CSI report is meant
to be used at the transmitter, not by the receiver hardware directly.

Instead, I think it makes more sense to model the quantization error
based on where it actually comes from, at the receiver ADC.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>