netlink: Use connector ID from out-of-tree range
This allows the iwlwifi module to be rebuilt for an existing kernel
and installed in the module 'updates' directory, without needing to
rebuild and install the entire kernel.

Also with this change, the Makefile no longer refers to an Ubuntu-
specific path to the kernel headers, allowing the netlink tools to
be built on other Linux distributions.

Signed-off-by: David Ward <david.ward@ll.mit.edu>